rename to simple_key_value_store
