+++ Cargo.toml
@@ -34,9 +34,3 @@ name = "routing"
 path = "examples/routing.rs"
 test = true
 bench = false
-
-[[example]]
-name = "routing_new"
-path = "examples/routing_new.rs"
-test = true
-bench = false
